fix height stats feature and move to fe folder
